auth: export handleAuthEvent, isAuthMatching
