+++ libs/langgraph-api/src/auth/custom.mts
@@ -1,13 +1,10 @@
-import type { AuthEventValueMap } from "@langchain/langgraph-sdk/auth";
 import type { MiddlewareHandler } from "hono";
 
 import {
-  authorize,
   authenticate,
   isAuthRegistered,
   isStudioAuthDisabled,
   type AuthContext,
-  type AuthFilters,
 } from "./index.mjs";
 
 declare module "hono" {
@@ -15,50 +12,6 @@ declare module "hono" {
     auth?: AuthContext | undefined;
   }
 }
-
-export function isAuthMatching(
-  metadata: Record<string, unknown> | undefined,
-  filters: AuthFilters
-) {
-  if (filters == null) return true;
-  for (const [key, value] of Object.entries(filters)) {
-    if (typeof value === "object" && value != null) {
-      if (value.$eq) {
-        if (metadata?.[key] !== value.$eq) return false;
-      } else if (value.$contains) {
-        if (
-          !Array.isArray(metadata?.[key]) ||
-          !metadata?.[key].includes(value.$contains)
-        ) {
-          return false;
-        }
-      }
-    } else {
-      if (metadata?.[key] !== value) return false;
-    }
-  }
-
-  return true;
-}
-
-export const handleAuthEvent = async <T extends keyof AuthEventValueMap>(
-  context: AuthContext | undefined,
-  event: T,
-  value: AuthEventValueMap[T]
-): Promise<[AuthFilters | undefined, value: AuthEventValueMap[T]]> => {
-  const [resource, action] = event.split(":");
-  const result = await authorize({
-    resource,
-    action,
-    context,
-    value,
-  });
-
-  return [result.filters, result.value] as [
-    AuthFilters | undefined,
-    value: AuthEventValueMap[T]
-  ];
-};
 
 const STUDIO_USER = {
   kind: "StudioUser",
